backend/requirements.txt
Django==3.2.25
djangorestframework==3.15.1
psycopg2-binary==2.9.6